+++ packages/appsync-modelgen-plugin/src/__tests__/visitors/appsync-visitor.test.ts
@@ -680,6 +680,21 @@ describe('AppSyncModelVisitor', () => {
       const updatedAtField = postFields.find(field => field.name === 'updatedOn');
       expect(updatedAtField).toMatchObject(updatedAtFieldObj);
     });
+    it('should not generate timestamp fields if "timestamps:null" is defined in @model', () => {
+      const schema = /* GraphQL */ `
+        type Post @model(timestamps: null) {
+          id: ID!
+        }
+      `;
+      const visitor = createAndGenerateVisitor(schema);
+      expect(visitor.models.Post).toBeDefined();
+
+      const postFields = visitor.models.Post.fields;
+      const createdAtField = postFields.find(field => field.name === 'createdAt');
+      expect(createdAtField).not.toBeDefined();
+      const updatedAtField = postFields.find(field => field.name === 'updatedAt');
+      expect(updatedAtField).not.toBeDefined();
+    });
   });
 
   describe('manyToMany testing', () => {
@@ -897,4 +912,29 @@ describe('AppSyncModelVisitor', () => {
       expect(visitor.models.ModelB.fields[1].directives[0].arguments.indexName).toEqual('byModelB');
     });
   });
+
+  describe('Graphql V2 fix tests for multiple has many relations of only one model type', () => {
+    const schema = /* GraphQL*/ `
+      type Registration @model {
+        id: ID! @primaryKey
+        meetingId: ID @index(name: "byMeeting", sortKeyFields: ["attendeeId"])
+        meeting: Meeting! @belongsTo(fields: ["meetingId"])
+        attendeeId: ID @index(name: "byAttendee", sortKeyFields: ["meetingId"])
+        attendee: Attendee! @belongsTo(fields: ["attendeeId"])
+      }
+      type Meeting @model {
+        id: ID! @primaryKey
+        title: String!
+        attendees: [Registration] @hasMany(indexName: "byMeeting", fields: ["id"])
+      }
+      
+      type Attendee @model {
+        id: ID! @primaryKey
+        meetings: [Registration] @hasMany(indexName: "byAttendee", fields: ["id"])
+      }
+    `;
+    it(`should not throw error when processing models`, () => {
+      expect(() => createAndGeneratePipelinedTransformerVisitor(schema)).not.toThrow();
+    });
+  })
 });