+++ packages/appsync-modelgen-plugin/src/visitors/appsync-typescript-visitor.ts
@@ -34,7 +34,9 @@ export class AppSyncModelTypeScriptVisitor<
   ];
 
   generate(): string {
-    this.processDirectives();
+    // TODO: Remove us, leaving in to be explicit on why this flag is here.
+    const shouldUseModelNameFieldInHasManyAndBelongsTo = false;
+    this.processDirectives(shouldUseModelNameFieldInHasManyAndBelongsTo);
     const imports = this.generateImports();
     const enumDeclarations = Object.values(this.enumMap)
       .map(enumObj => this.generateEnumDeclarations(enumObj))
@@ -244,10 +246,12 @@ export class AppSyncModelTypeScriptVisitor<
 
   protected getNativeType(field: CodeGenField): string {
     const typeName = field.type;
+    const isNullable = field.isList ? field.isListNullable : field.isNullable;
+    const nullableTypeUnion = isNullable ? ' | null' : '';
     if (this.isModelType(field)) {
       const modelType = this.modelMap[typeName];
       const typeNameStr = this.generateModelTypeDeclarationName(modelType);
-      return field.isList ? this.getListType(typeNameStr, field) : typeNameStr;
+      return (field.isList ? this.getListType(typeNameStr, field) : typeNameStr) + nullableTypeUnion;
     }
 
     let nativeType = super.getNativeType(field);
@@ -255,6 +259,8 @@ export class AppSyncModelTypeScriptVisitor<
     if (this.isEnumType(field)) {
       nativeType = `${nativeType} | keyof typeof ${this.getEnumName(this.enumMap[typeName])}`;
     }
+
+    nativeType = nativeType + nullableTypeUnion;
 
     return nativeType;
   }